pyproject.toml
[project]
name = "diff-shakespeare"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=2.0.0",
    "gradio>=4.0.0",
    "huggingface-hub>=0.17.0",
    "ipykernel>=6.25.2",
    "python-dotenv>=1.2.2",
    "tiktoken>=0.6.0",
    "torch>=2.0.0",
]
